+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next"
 import { Geist, Geist_Mono } from "next/font/google"
 import "./globals.css"
-import { Toaster } from "@/components/ui/sonner"
-import { QueryProvider } from "@/components/providers/query-provider"
+import { Toaster } from "@/components/ui/Sonner"
+import { QueryProvider } from "@/components/providers/QueryProvider"
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
