+++ src/app/official/[brand]/not-found.tsx
@@ -3,9 +3,9 @@
  */
 
 import Link from "next/link"
-import { MainLayout } from "@/components/layout/main-layout"
-import { Button } from "@/components/ui/button"
-import { Card } from "@/components/ui/card"
+import { MainLayout } from "@/components/layout/MainLayout"
+import { Button } from "@/components/ui/Button"
+import { Card } from "@/components/ui/Card"
 import { AlertTriangleIcon, HomeIcon } from "lucide-react"
 
 export default function BrandNotFound() {
