+++ src/components/layout/MainLayout.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from "./sidebar"
+import { Sidebar } from "./Sidebar"
 
 interface MainLayoutProps {
   children: React.ReactNode
